Evo 18.5 Upload
Reverted back to 16.8 target selection code and 16.8 setServo subroutine
usage.  Fixed logo display issue on startup, as well as moved and
modified the servo centering step, so it centered the servo and then
released the servo.  This resulted in a reduction of current draw by the
servo during motor startup.  Once started, the servo could be re-engaged
from the "center" position.  This code also uses the GY-86 board, which
requires the MPU-6050 to be set to I2C Passthrough, in order to directly
communicate with the compass.

A physical change to the PCB was to remove and bypass the 300mA
resettable fuse and remove the protection diode.